% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate.R
\name{allocate}
\alias{allocate}
\title{Allocates position based on lat-long coordinates to a polygon geometry}
\usage{
allocate(coords, geometries)
}
\arguments{
\item{coords}{\code{sf} object containing a data frame with coordinate data as point geometries}

\item{geometries}{\code{sf} object containing geometries to map to. Should contain at least two named columns: \code{"id"} and \code{"geometry"}}

\item{quadrant}{\code{Boolean} denoting whether the geometry {sf} is for quadrants or not. Defaults to \code{FALSE}}
}
\value{
\code{sf} object containing mapping
}
\description{
Allocates position based on lat-long coordinates to a polygon geometry
}
\author{
Trent Henderson
}
