% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve.R
\name{dissolve}
\alias{dissolve}
\title{Dissolve boundaries from multiply polygons into a single geometric polygon}
\usage{
dissolve(geometries, .group = NULL)
}
\arguments{
\item{geometries}{\code{sf} object contain the geometry}

\item{.group}{\code{character} denoting the column name of the grouping variable to aggregate over (if applicable). Defaults to \code{NULL}. If \code{NULL}, all boundaries in \code{data} will be dissolved into a single polygon geometry}
}
\value{
\code{sf} object
}
\description{
Dissolve boundaries from multiply polygons into a single geometric polygon
}
\author{
Trent Henderson
}
