% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation.R
\name{expectation}
\alias{expectation}
\title{Draw a sample point from a multivariate Gaussian parameterised by longitude and latitude and ensure it is within geometric boundaries}
\usage{
expectation(
  longitude,
  latitude,
  geometry = NULL,
  seed = 123,
  return_cov = FALSE
)
}
\arguments{
\item{longitude}{\code{numeric} vector containing longitude data}

\item{latitude}{\code{numeric} vector containing latitude data}

\item{seed}{\code{integer} denoting the fix for R's pseudo-random number generator. Defaults to \code{123}}

\item{return_cov}{\code{Boolean} denoting whether to return the mean vector and covariance matrix in addition to sampled point. Defaults to \code{FALSE}. Returns a \code{list} object if \code{TRUE}}

\item{geometries}{\code{sf} object containing geometric boundary to check point is within. Defaults to \code{NULL}}
}
\value{
\code{sf} object containing the sampled point
}
\description{
Draw a sample point from a multivariate Gaussian parameterised by longitude and latitude and ensure it is within geometric boundaries
}
\author{
Trent Henderson
}
