% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside.R
\name{inside}
\alias{inside}
\title{Map smaller polygons in one object to larger polygons in another by area intersection}
\usage{
inside(A, B)
}
\arguments{
\item{A}{\code{sf} object containing the larger polygons}

\item{B}{\code{sf} object containing the smaller polygons to be mapped}
}
\value{
\code{sf} object containing the smaller geometry data but with mapping information
}
\description{
Map smaller polygons in one object to larger polygons in another by area intersection
}
\author{
Trent Henderson
}
