% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.spatialdensity}
\alias{plot.spatialdensity}
\title{Plots geospatial kernel density in ggplot2}
\usage{
\method{plot}{spatialdensity}(x, ...)
}
\arguments{
\item{...}{arguments to be passed to methods}

\item{data}{\code{sf} object containing a data frame of the data to plot}

\item{geometries}{\code{sf} object containing geometries}
}
\value{
\code{ggplot} object containing the graphic
}
\description{
Plots geospatial kernel density in ggplot2
}
\author{
Trent Henderson
}
