% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Divide an sf object polygon into some number of roughly equally-sized parts and calculate the centroid coordinates}
\usage{
segment(sf_poly, n_zones = 25, n_points = 10000, seed = 123)
}
\arguments{
\item{sf_poly}{\code{sf} object contain the geometry}

\item{n_zones}{\code{integer} denoting the number of zones to divide the geometry into. Defaults to \code{25}}

\item{n_points}{\code{integer} denoting the number of points to use to construct the clustering algorithm. Defaults to \code{1e4}}

\item{seed}{\code{integer} denoting the fix for R's pseudo-random number generator. Defaults to \code{123}}
}
\value{
\code{sf} object
}
\description{
Divide an sf object polygon into some number of roughly equally-sized parts and calculate the centroid coordinates
}
\references{
https://gis.stackexchange.com/questions/375345/dividing-polygon-into-parts-which-have-equal-area-using-r

https://gis.stackexchange.com/questions/321021/splitting-polygon-into-equal-area-polygons-using-qgis

https://blog.cleverelephant.ca/2018/06/polygon-splitting.html

https://www.khetarpal.org/polygon-splitting/
}
\author{
Trent Henderson
}
