% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialkernel.R
\name{spatialkernel}
\alias{spatialkernel}
\title{Computes smoothed spatial kernel density estimation using data and geometry information}
\usage{
spatialkernel(data, geometries, ...)
}
\arguments{
\item{data}{\code{sf} object containing a data frame of the data to plot}

\item{geometries}{\code{sf} object containing geometries}

\item{...}{arguments to be passed to methods}
}
\value{
\code{spatialdensity} object containing the estimated spatial kernel density information and geometric information
}
\description{
Computes smoothed spatial kernel density estimation using data and geometry information
}
\references{
https://stackoverflow.com/questions/68643517/smoothed-density-maps-for-points-in-using-sf-within-a-given-boundary-in-r
}
\author{
Trent Henderson
}
